add prefix to logs

+++ infra/move-and-publish-build.sh
@@ -6,9 +6,14 @@ current_branch=$(git rev-parse --abbrev-ref HEAD)
 
 echo moving and commiting at $output_path
 
-git fetch
-git switch -C ${branch_name} origin/${branch_name}
+git fetch && \
+git switch -C ${branch_name} origin/${branch_name} && \
 git pull
+
+if [ $? -gt 0 ]; then
+	echo "MNP: something went wrong, exiting"
+	exit $?
+fi
 
 rm -rf ./dist && \
 mv -v -f ./$output_path/dist ./ && \
